chore: split printers
